Add usage of click library for command line argument parsing
