+++ requirements.txt
@@ -1,3 +1,4 @@
+click>=7.1.2
 inquirer>=2.7.0
 requests>=2.24.0
 ruamel.yaml>=0.16.12
